Update main.cpp
Changed the previous software watchdog timer to the built in avr watchdog timer that could handle resets and error checking. 
Refined the loop to accommodate the changes and removed the for loops in setting the pins as well as adding the wdt (watchdog timer) avr function setup into the setup function of the code. 
NOT TESTED.